And then what I'm going to do is I am going to call another use effect. ln 19
This filtered monster's call.

And here I'm going to say this is new filtered monsters.

And once we filter through and we have that new monster, then we are going to say set filtered monsters

to the new filtered monsters.

And now filtered monsters here is always going to contain the values we want.

And what I'm going to do is I'm going to say, I want you to filter through these monsters whenever

either the monsters array changes or whenever the search field changes.

so video jätk # 60 osast.