Most of the time you're just going to use props, and this props is the exact same as this props.
It's an object that has all of the different parameters and properties essentially that are being passed

into this component.

It's just that instead of this props, it's going to be the first parameter that gets passed into this

functional component.

Similarly, because we know that this will always be the first parameter, we actually don't need to

do the structuring here.

We can do the structuring right inside of the parameter itself, like so.